add custom endpoint attribute
