+++ scaleway/helpers_mnq_queue.go
@@ -3,17 +3,113 @@ package scaleway
 import (
 	"context"
 	"fmt"
+	"net/http"
+	"os"
 	"regexp"
 	"strconv"
 	"strings"
 
 	"github.com/aws/aws-sdk-go/aws"
+	"github.com/aws/aws-sdk-go/aws/credentials"
+	"github.com/aws/aws-sdk-go/aws/session"
 	"github.com/aws/aws-sdk-go/service/sqs"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/id"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
+	"github.com/nats-io/nats.go"
 	mnq "github.com/scaleway/scaleway-sdk-go/api/mnq/v1alpha1"
 	"github.com/scaleway/scaleway-sdk-go/scw"
 )
+
+func SQSClientWithRegion(d *schema.ResourceData, m interface{}) (*sqs.SQS, scw.Region, error) {
+	meta := m.(*Meta)
+	region, err := extractRegion(d, meta)
+	if err != nil {
+		return nil, "", err
+	}
+
+	if _, ok := d.GetOk("sqs"); !ok {
+		return nil, "", fmt.Errorf("sqs access_key and secret_key are required")
+	}
+
+	endpoint := d.Get("sqs.0.endpoint").(string)
+	accessKey := d.Get("sqs.0.access_key").(string)
+	secretKey := d.Get("sqs.0.secret_key").(string)
+
+	sqsClient, err := newSQSClient(meta.httpClient, region.String(), endpoint, accessKey, secretKey)
+	if err != nil {
+		return nil, "", err
+	}
+
+	return sqsClient, region, err
+}
+
+func newSQSClient(httpClient *http.Client, region string, endpoint string, accessKey string, secretKey string) (*sqs.SQS, error) {
+	config := &aws.Config{}
+	config.WithRegion(region)
+	config.WithCredentials(credentials.NewStaticCredentials(accessKey, secretKey, ""))
+	config.WithEndpoint(strings.ReplaceAll(endpoint, "{region}", region))
+	config.WithHTTPClient(httpClient)
+	if strings.ToLower(os.Getenv("TF_LOG")) == "debug" {
+		config.WithLogLevel(aws.LogDebugWithHTTPBody)
+	}
+
+	s, err := session.NewSession(config)
+	if err != nil {
+		return nil, err
+	}
+	return sqs.New(s), nil
+}
+
+func NATSClientWithRegion(d *schema.ResourceData, m interface{}) (nats.JetStreamContext, scw.Region, error) { //nolint:ireturn
+	meta := m.(*Meta)
+	region, err := extractRegion(d, meta)
+	if err != nil {
+		return nil, "", err
+	}
+
+	if _, ok := d.GetOk("nats"); !ok {
+		return nil, "", fmt.Errorf("nats credentials are required")
+	}
+
+	endpoint := d.Get("nats.0.endpoint").(string)
+	credentials := d.Get("nats.0.credentials").(string)
+	js, err := newNATSJetStreamClient(region.String(), endpoint, credentials)
+	if err != nil {
+		return nil, "", err
+	}
+
+	return js, region, err
+}
+
+func newNATSJetStreamClient(region string, endpoint string, credentials string) (nats.JetStreamContext, error) { //nolint:ireturn
+	jwt, seed, err := splitNATSJWTAndSeed(credentials)
+	if err != nil {
+		return nil, err
+	}
+
+	nc, err := nats.Connect(strings.ReplaceAll(endpoint, "{region}", region), nats.UserJWTAndSeed(jwt, seed))
+	if err != nil {
+		return nil, err
+	}
+
+	js, err := nc.JetStream()
+	if err != nil {
+		return nil, err
+	}
+
+	return js, nil
+}
+
+func splitNATSJWTAndSeed(credentials string) (string, string, error) {
+	lines := strings.Split(credentials, "\n")
+	if len(lines) < 6 {
+		return "", "", fmt.Errorf("invalid credentials format")
+	}
+
+	jwt := lines[1]
+	seed := lines[4]
+	return jwt, seed, nil
+}
 
 const SQSFIFOQueueNameSuffix = ".fifo"
